Added CustomBubbleValueToTextConverter
This adds a custom bubble text converter to create the possibility to map a specific progress value to a different value that is displayed to the user. Also you can now add something like "2h" as your bubble text. To archive this I changed the following:

added DefaultConverter to config
changed every bubble text to be set through converter